UHF-6017: move mobilemenu to separate file. fix transition scroll positions
