+++ package.json
@@ -42,6 +42,7 @@
     "husky": "^8.0.0",
     "joi": "^17.4.0",
     "lint-staged": "^13.0.0",
+    "lodash": "^4.17.21",
     "md5": "^2.3.0",
     "mini-css-extract-plugin": "^1.6.2",
     "minivents": "^2.2.1",
