+++ src/scss/06_components/navigation/_mobile-navigation.scss
@@ -424,16 +424,26 @@ $-cssmenu-bottom-padding: 80px; // Reserve space for mobile browser untappable-a
 * JS menu styles (reuse CSS menu classes where possible, remove unused when done
 */
 
-.jsmenu {
-  display: flex;
-  height:80vh;
+.jsmenu__drawer {
+  background:#ddd;
   overflow-x:hidden;
-  position:relative;
-  width:1000rem
+  position:absolute;
+
 }
+
+.jsmenu__panels  {
+  background:#ddd;
+  height:100vh;
+  position:relative;
+  width:100vw;
+  width:1000rem;
+}
+
 
 .jsmenu__panel {
   background-color: white;
+  display:flex;
+  flex-direction: column;
   min-height:100vh;
   position:absolute;
   transition-property: all;
@@ -459,25 +469,49 @@ $-cssmenu-bottom-padding: 80px; // Reserve space for mobile browser untappable-a
   transition-timing-function: ease-out;
 }
 
-
 .jsmenu__panel.jsmenu__panel--visible-up {
-  opacity:0.3;
+  opacity:0.1;
   transform:translateX(100vw);
   transition-property: all;
   transition-timing-function: ease-in;
 }
 
 .jsmenu__panel.jsmenu__panel--visible-down {
-  opacity:0.3;
+  opacity:0.1;
   transform:translateX(-100vw);
   transition-property: all;
   transition-timing-function: ease-in;
 }
 
+.jsmenu__panel-body {
+  min-height: 60vh;
+}
+
+.jsmenu__panel-footer {
+  background:#ddd;
+  flex-grow: 1;
+  min-height:10rem;
+  padding-left:$spacing;
+  padding-top:$spacing-and-half;
+}
+
+
+.jsmenu__panel--current .jsmenu__panel-footer {
+  // for transition purposes
+  min-height:100vh;
+}
+
 .jsmenu__language  {
   background-color: #ddd;
   height:4rem;
+  position: relative;
 }
+
+// .jsmenu__language::after {
+//   position: absolute;
+//   right:-2rem;
+//   @include pseudo-icon('angle-down', $spacing-and-half, black);
+// }
 
 .jsmenu__items {
   background-color: white;
@@ -493,9 +527,9 @@ $-cssmenu-bottom-padding: 80px; // Reserve space for mobile browser untappable-a
   display:flex;
   font-size:125%;
   gap:12px;
-  height: 3rem;
   margin:0;
-  padding-left:1rem;
+  min-height: $spacing-triple;
+  padding-left:$spacing;
 }
 
 .jsmenu__itemlink {
@@ -510,9 +544,23 @@ $-cssmenu-bottom-padding: 80px; // Reserve space for mobile browser untappable-a
   flex-grow:0;
   flex-shrink: 0;
   height:3rem;
-  width: 4rem;
+  width: 3rem;
 }
 
-.jsmenu__panel-footer {
-  background:#ddd;
+.jsmenu__button--forward::before {
+  @include pseudo-icon('angle-right', $spacing-and-half, black);
+  // @include rotating-icon-transition;
 }
+
+.jsmenu__button--back {
+  background:none;
+  border:0;
+  height:3rem;
+}
+
+.jsmenu__panel:not(:first-child) .jsmenu__button--back::before {
+  @include pseudo-icon('angle-left', $spacing-and-half, black);
+  // @include rotating-icon-transition;
+}
+
+
