+++ templates/navigation/menu--mobile.html.twig
@@ -24,8 +24,9 @@
   We call a macro which calls itself to render the full tree.
   @see https://twig.symfony.com/doc/1.x/tags/macro.html
 #}
-<div id="jsmenu" class="jsmenu">
-</div>
+<nav id="jsmenu" class="jsmenu__drawer">
+  <div id="jsmenu__panels" class="jsmenu__panels"/>
+</nav>
 {# {{ menus.menu_links(items, attributes, 0) }} #}
