+++ webpack.config.js
@@ -31,7 +31,7 @@ const Entries = () => {
   ];
 
   glob.sync(pattern, {ignore: ignore}).map((item) => {
-    entries[path.parse(item).name] = item }
+    entries[path.parse(item).name] = item; }
   );
   return entries;
 };
@@ -169,7 +169,7 @@ module.exports = (env, argv) => {
         minimizer: [
           new TerserPlugin({
             terserOptions: {
-              ecma: 2015,
+              ecma: 2020,
               format: {
                 comments: false,
               },
